queue: always fill in sq->array on io_uring_submit()
io_uring_submit() is a last commit point when a caller has to be sure
that all prepared sqes eventually lands to the kernel.

If SQ thread is running it is quite possible that khead != ktail, thus
charging of sq->array is skipped, which leads to IO hang.

Nasty hang is well reproduced when submitter and completion harverster
are different threads.

Signed-off-by: Roman Penyaev <rpenyaev@suse.de>
Signed-off-by: Jens Axboe <axboe@kernel.dk>