Add libaio wrapper
We can wrap the aio interface with io_uring.

Very basic wrapper, but seems to work. To test with fio, for example,
just change the following in configure where it adds libaio:

    LIBS="-laio $LIBS"

to

    LIBS="-luring $LIBS"

which will cause fio to link with just liburing and use the compat
wrappers instead of actual aio.

This should be good enough for testing, though I'm sure there corner
cases to iron out. Most notably this doesn't support for aio user
mapped ring for completions. We could add support for that, but that
requires kernel side additions too. Totally doable, but not sure how
far to take this instead of just converting folks to use io_uring
instead.

Signed-off-by: Jens Axboe <axboe@kernel.dk>