Add test for overflow of timeout request's sequence
Before 5da0fb1ab34c ("io_uring: consider the overflow of sequence for
timeout req"). We can meet some situation like below:

1. setup
2. prepare 4 timeout req which expected count is 1,1,2,UINT_MAX, and the
sequence of this 4 requests will be 1,2,4,2, this 4 requests will not
lead the change of cached_cq_tail and sq_dropped until the timeout
really happened. So the tail_index in io_timeout will still be 0.
3. based on the above and before this patch, the order of timeout_list
will be req1->req2->req4->req3, which the right order should be
req1->req2->req3->req4.
4. setup two nop requests. And the timeout requests will return
correctly with the patch.

Add this testcase to cover it.

Signed-off-by: yangerkun <yangerkun@huawei.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>