Fix io_uring_get_sqe() for IORING_SETUP_SQPOLL
If we use kernel side polling, we can't use the flush done by
io_uring_submit() as the source of truth for whether or not we
can grab more SQE entries. We need to use the kernel side for that,
as the SQPOLL thread could not have caught up yet.

Signed-off-by: Jens Axboe <axboe@kernel.dk>