Don't let io_uring_wait_cqes() submit IO
It's a wait interface, if we let it it also touch the SQ side of things,
then it's not possible to use sanely in a situation where someone has
a separate thread for completions.

Note that this may require application changes, if they are currently
relying on io_uring_wait_cqes() on also submitting IO. The risk should
be minimal however, since it is a wait-for-completions interface.

Signed-off-by: Jens Axboe <axboe@kernel.dk>