Split out system call bits from library interface
Rename the io_uring_* system calls to __sys_io_uring_* and make them
locals.

We only have the system calls because they are not in glibc yet, and
it's somewhat confusing that they share the same namespace as the
library functions. With this change, any exported io_uring_* function
is a library function.

Signed-off-by: Jens Axboe <axboe@kernel.dk>