add helper functions to verify io_uring supported opcodes
It is common for an application using an ever-evolving interface to want
to inquire about the presence of certain functionality it plans to use.

Information about opcodes is stored in a io_uring_probe structure. There
is usually some boilerplate involved in initializing one, and then using
it to check if it is enabled.

This patch adds two new helper functions: one that returns a pointer to
a io_uring_probe (or NULL if probe is not available), and another one
that given a probe checks if the opcode is supported.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
CC: Avi Kivity <avi@scylladb.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>