liburing.map: Fix LIBURING_0.3 shared library version node
This version node is disconnected from LIBURING_0.2, and contains a
merged duplication of the LIBURING_0.1 and LIBURING_0.2 version nodes,
plus few local entries.

Remove its contents and link it to the LIBURING_0.2, as it will inherit
everything there in addition to the catchall local in LIBURING_0.1.

This leaves an empty version node around, and while it is a bit of an
oddity and it might be somewhat safe to remove, it could still break
users. So we leave it behind as a representation of what went on in
that release.

Signed-off-by: Guillem Jover <guillem@hadrons.org>