test: Use __NR_mmap2 when __NR_mmap is not defined
On some new 32-bit architectures (such as armel and armhf), __NR_mmap
is not defined, but we can safely use __NR_mmap2 as we are passing an
offset of 0, so the different bases for that argument do not apply.

We need to stop defining __NR_mmap in one of the unit tests to be able
to properly fallback to the __NR_mmap2 case.

Signed-off-by: Guillem Jover <guillem@hadrons.org>