io_uring_get_sqe: always use khead
If we submit a batch of requests via io_uring_submit() but get some
kind of error (like -EBUSY), the internal state has been flushed to
the SQ ring. We can then get into a situation where we think have more
than ring-size number of requests available, which can cause duplicate
submissions of requests.

Should be safe to always use the shared state for this, and it avoids
repeat flushing bumping the size.

Reported-by: Roman Gershman <romange@gmail.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>