Bump major version to 2
A few reasons for this:

- Allow us to pad out the exposed structures so we've got some room
  left to make changes without breaking compatability.
- There seems to be a bit of a mixup on versions, 0.x vs 1.x. We called
  them 0.x, but the major versionw as 1 already.

While in there, also correct the mistake from commit:

29768113015a ("Add io_uring_sqring_wait()")

where the symbol was added to 0.7, but should have been 0.8 instead as
pointed out by Guillem Jover.

Note that 2.x is fully compatible with 1.x, the structures are just
padded out. All that is needed is a recompile of applications, and that's
not _actually_ needed as the padding is all at the end and we don't use
any extra space right now anyway.

Signed-off-by: Jens Axboe <axboe@kernel.dk>