test: Pass an array to run_tests instead of an undefined variable
The tests explicitly check the number of arguments and will skip them in
case they are passed one if they are not supposed to be handling it.

Do not pass an undefined variable, as that means we cannot quote it,
which is not future-proof. Instead use an array and append the dev
variable to it, if it is not defined.

Warned-by: shellcheck
Signed-off-by: Guillem Jover <guillem@hadrons.org>