test/file-verify: add read verification test
This test attempts to exercise the various paths that a read can take.
We try and trigger incremental reads and loops, by removing page cache
parts of a cached file for reading. In essence, these are the two
types of iterator conditions we want to hit:

1) We get -EAGAIN immediately. This is common for buffered IO when
   the range isn't in page cache.
2) We get a short read, again common for buffered read.

Internally, we have three different mechanisms we want to exercise,
as they can behave a bit differently:

1) Using non-vectored read
1) Using non-vectored registered buffer read
3) Using vectored read, segments that fit in UIO_FASTIOV
4) Using vectored read, segments larger than UIO_FASTIOV

We try and accomplish this by doing reads that hit any of these three
conditions, and then we remove page cache in chunks from a file from
either the front, middle, or end of the file. We do go through all
of these for both buffered and O_DIRECT reads.

Signed-off-by: Jens Axboe <axboe@kernel.dk>