Added support for CMake build system
Created CMakeLists.txt and updated .gitignore accordingly. Having
support for CMake enables liburing  to be added as a dependency via the
FetchContent interface, so that other projects build and link against
liburing on supported kernel versions even if no system installation is
present. This can be useful on HPC systems or other shared clusters
where the approval process for system-wide installation is slow. It also
makes it easier for users to test and compare more recent versions of
the library to an existing system installation.

Signed-off-by: Alecto Perez <perez.cs@pm.me>