src/Makefile: Always use `-fpic -fPIC` flag for nolibc build
Bob reports linking time error for nolibc build:
```
  relocation R_X86_64_PC32 against symbol `memset' can not be \
  used when making a shared object; recompile with -fPIC
```
When we statically link liburing.a into a shared object, we need to
make sure it's built with `-fpic -fPIC` flag.

Note:
liburing.a can be used to create a new shared object (*.so) or a
non shared object ELF. When we use it to create a shared object,
it needs to be compiled with the above flag.

Always using this flag for nolibc build even if the target is not
a shared object ELF is still fine.

Link: https://github.com/axboe/liburing/issues/443#issuecomment-975493681
Fixes: c7d03dc2c123bc2ee85582d8f922acd268d6ac78 ("configure: Add `CONFIG_NOLIBC` variable and macro")
Fixes: f48ee3168cdc325233825603269f304d348d323c ("Add nolibc build support")
Reported-by: Bob Chen <beef9999@qq.com>
Tested-by: Bob Chen <beef9999@qq.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>