liburing.h: don't truncate timeout_update/remove user data value
If we're running on 32-bit, with the cast that io_uring_prep_rw() does,
we lose the top 32 bits of the user_data for the timeout remove and
update commands. If the application also uses a full 64-bit value for
the original timeout user_data, then we cannot find the timeout to
update.

Link: https://github.com/axboe/liburing/issues/490
Signed-off-by: Rémi Bernon <rbernon@codeweavers.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>