configure: Fix clang-12 warning `-Wdeprecated`
```
clang-12: warning: treating 'c' input as 'c++' when in C++ mode, this \
behvior is deprecated [-Wdeprecated]
```

The target compile is C++ but the file extension is ".c". Got the
above warning from clang. Fix it by creating a temporary file with
".cpp" extension and use it.

Introduced by commit d9f953770708ad ("configure: test for presence of
C++ compiler").

Fixes: d9f953770708ad6893140dc2d11258223ba7abfd ("configure: test for presence of C++ compiler")
Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>