Add preliminary support for using a registered ring fd
Add a specific ->enter_ring_fd member to the ring, and just set it to
the normal ring fd at init time.

Add an internal flag to track whether or not we're actually using a
registered fd for the ring itself, and set the appropriate
io_uring_enter(2) if so.

Signed-off-by: Jens Axboe <axboe@kernel.dk>