Add test for multiple concurrent accepts
Add tests for accept that queues multiple accepts and then ensures the
correct things happen.

Check that when connections arrive one at a time that only one CQE is
posted (in blocking and nonblocking sockets), as well as make sure that
closing the accept socket & cancellation all work as expected.

This relies on a kernel with [1] for the tests to pass.

[1]: https://lore.kernel.org/io-uring/20220325093755.4123343-1-dylany@fb.com/

Signed-off-by: Dylan Yudaken <dylany@fb.com>
Link: https://lore.kernel.org/r/20220325094013.4132496-1-dylany@fb.com
Signed-off-by: Jens Axboe <axboe@kernel.dk>