io_uring.h: add IORING_RECVSEND_POLL_FIRST
This flag can be used by send/sendmsg or recv/recvmsg to tell io_uring
that it expects the socket to not be immediately ready for send or
recv. For that case, it's worthless trying to do IO first, it's better
to just arm poll internally and let that drive the first attempt.

Can be used in conjunction with the IORING_CQE_F_SOCK_NONEMPTY feedback
from the kernel, or independently.

Signed-off-by: Jens Axboe <axboe@kernel.dk>