man: add note on vectored reads and writes
Applications tend to use vectored IO even for a single segment IO
request, but this is less efficient than a non-vectored one. This is
true for the normal readv(2)/writev(2) and related system calls as
well, but it doesn't seem that this is common knowledge.

In short, unless you have more than 1 segment for an IO request, use
the non-vectored version!

Signed-off-by: Jens Axboe <axboe@kernel.dk>