tests: correctly exit with failure in a looped test
A common test pattern in this project is to test a couple related things
in a single test binary, and return failure on the first one that fails,
if any. Tracking subtest failures cannot be done well with simple exit()
statuses, though it can using something like TAP and parsing a generated
report.

However, this test simply set the value of `ret`, then proceeded to
another test. If the first failed and the second succeeded, we would log
a failure but then return a success.

Just return immediately on failure as is done elsewhere.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Link: https://lore.kernel.org/r/20220706034059.2817423-7-eschwartz93@gmail.com
Signed-off-by: Jens Axboe <axboe@kernel.dk>