Ensure LFS is used on 32-bit compiles
I noticed file-verify.t failures on 32-bit compiles, and it's due to
not using the correct 64-bit LFS sizing. This means off_t is 32-bit,
rather than the 64-bit we expect.

Just set _LARGEFILE_SOURCE and _FILE_OFFSET_BITS=64 unconditionally,
and remove it from the two tests where we already used it.

Signed-off-by: Jens Axboe <axboe@kernel.dk>