configure: Always enable `CONFIG_NOLIBC` if the arch is supported
Currently, the default liburing compilation uses libc as its dependency.
liburing doesn't depend on libc when it's compiled on x86-64, x86
(32-bit), and aarch64. There is no benefit to having libc.so linked to
liburing.so on those architectures. Always enable CONFIG_NOLBIC if the
arch is supported. If the architecture is not supported, fallback to
libc.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Co-authored-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Link: https://lore.kernel.org/r/20230106155202.558533-3-ammar.faizi@intel.com
Signed-off-by: Jens Axboe <axboe@kernel.dk>