multicqes_drain: make trigger event wait before reading
trigger_event is used to generate CQEs on the poll requests. However there
is a race if that poll request is running asynchronously, where the
read_pipe will complete before the poll is run, and the poll result will
be that there is no data ready.

Instead sleep and force an io_uring_get_events in order to give the poll a
chance to run before reading from the pipe.

Signed-off-by: Dylan Yudaken <dylany@meta.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>