test/connect: skip if tcp_syncookies isn't enable
This reliably triggers:

TCP: request_sock_TCP: Possible SYN flooding on port 127.0.0.1:26152. Dropping request.

and we then drop the first request and the test just sits there waiting
for it to connect. Ideally we just want the first connect to succeed and
second one to be timed out, but not quite sure how to accomplish that.
If someone has an idea to make that work, please do send an update to the
test!

Signed-off-by: Jens Axboe <axboe@kernel.dk>