test/poll-race-mshot: fix missing init of buffer ring
It's critical to initialize the ring before use, obviously... As is also
documented.

Outside of that, intialize the buffer written. This doesn't really
matter, but it's a nice thing to do.

Use the appropriate helpers to add buffers to the ring, rather than
maintain a private mask.

Finally, free the ring and buffers when done. This usually doesn't matter
in test cases, but we are calling these thousands of times. Better to
free the memory than bloat the usage.

Signed-off-by: Jens Axboe <axboe@kernel.dk>