test/file-verify: add dcache sync for parisc
If registered buffers are used, the application may need to flush
the dcache for the buffer if they are reused. This should probably
be a liburing helper, but as this only really affects parisc as far
as I can tell, we'll just stuff it in the test case itself for now
and worry about that later, if needed.

Signed-off-by: Jens Axboe <axboe@kernel.dk>