Default to mmap'ed provided buffers for hppa
Architectures like parisc cannot guarantee that the kernel and userspace
virtual address map to the same cache color, when we have the application
allocate the ring and then the kernel mapping it.

Add a helper to use the buffer ring registration that has the kernel
allocate the memory, and then have the application mmap(2) it. With
that, we can have mmap pick an appropriate virtual address that so we
avoid any coloring mismatches.

Signed-off-by: Jens Axboe <axboe@kernel.dk>