Update io_uring_prep_fallocate() to use 64-bit length/offset
For 64-bit builds this was already the case, and for 32-bit builds with

-D_LARGEFILE_SOURCE -D_FILE_OFFSET_BITS=64

enabled. You could argue that without setting those you cannot work with
big files anyway, which is true, but it's somewhat inconvenient for
non C style languages.

Just make them 64-bit by default, we're already passing in 64-bits for
the kernel side anyway in the SQE.

Link: https://github.com/tokio-rs/io-uring/pull/127
Signed-off-by: Jens Axboe <axboe@kernel.dk>