recv-msgall: Fix undefined behavior in `recv_prep()`
The lifetime of `struct msghdr msg;` must be long enough until the CQE
is generated because the recvmsg operation will write to that storage. I
found this test segfault when compiling with -O0 optimization. This is
undefined behavior and may behave randomly.

Fix this by making the lifetime of `struct msghdr msg;` long enough.

Fixes: https://github.com/axboe/liburing/issues/855
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Link: https://lore.kernel.org/r/20230510143927.123170-2-ammarfaizi2@gnuweeb.org
Signed-off-by: Jens Axboe <axboe@kernel.dk>