setup: Avoid zeroing a freshly mapped MAP_ANONYMOUS page
The kernel promises that the memory returned by mmap with MAP_ANONYMOUS
is initialized to zero. We only need to zero a user-provided buffer.

We also only need to zero the memory we actually use from the
user-provided buffer, so that if the user puts several rings in one
allocation, we don't repeatedly re-zero the tail of the buffer past what
we use.

Signed-off-by: Josh Triplett <josh@joshtriplett.org>