setup: Use non-huge pages if sufficient for the queue sizes
Huge pages add friction and potential failures; they require system
configuration (nr_hugepages or nr_overcommit_hugepages) to permit
allocating them in the first place, and in the overcommit case they'll
fail if there isn't enough contiguous memory available to allocate one.

For small numbers of SQs and CQs, the queues fit entirely within one or
two normal-sized pages, so don't allocate huge pages in this case.

For 4096-byte pages, this allows a queue size of 64 to fit within two
normal page allocations, and a queue size of 32 to fit within one normal
page allocation. This should allow many users of io_uring to avoid
requiring huge pages.

To avoid tracking additional state for the size of the SQEs allocation,
rely on the munmap semantic of rounding up the requested unmap length to
the containing page.

Signed-off-by: Josh Triplett <josh@joshtriplett.org>