build: Disable stack protector unconditionally
At least when building with nolibc, we cannot use the stack protector
as we will end up with undefined references to __stack_chk_fail which
is provided by the libc. We need to add the -fno-stack-protector
unconditionally to the CFLAGS append override.

This seems to have been the intent behind the introduction of this
change in commit 2de98320d5b02951936fc0ab677dd01e4fb2a7a7.

Signed-off-by: Guillem Jover <guillem@hadrons.org>