Update sqe layout and prep helpers
Let's be smart about it and actually set aside sqe->futex_flags for
io_uring specific flags, and move the FUTEX2_* flags to sqe->fd instead.
This leaves us open to have tweaks internally for io_uring futex commands,
rather than need to add FUTEX2_ flags for that which may not at all make
any sense for the core futex2(2) syscalls.

This has bitten us in the bit previously and we've had to add a new
opcode rather than just add a flag, the most recent example of that is
the IORING_OP_READ_MULTISHOT which clearly could've just been a flag
for the READ/READV/READ_FIXED opcodes.

While doing so, make sure we have two sets of flags passed to the prep
helpers, so that we're future proof in that regard.

Signed-off-by: Jens Axboe <axboe@kernel.dk>