Link against libgcc based on output of -print-libgcc-file-name
GCC and clang have the -print-libgcc-file-name option to automatically
print out the correct compiler runtime library to link to. This can be
helpful in case the runtime library is named something other than libgcc
(i.e. on a system where only LLVM's compiler-rt is used), or if libgcc
is in a non-standard directory. If the option fails for whatever reason,
fall back to using "-lgcc".

Signed-off-by: Violet Purcell <vimproved@inventati.org>