examples/proxy: thorough cleaning and bug fixes
Generally not a huge fan of big changes like this, but the initial commit
was mostly done early so that someone else could have easy access to it.

This makes it work for a variety of more cases, and serializes sends
just in case we get them reordered on the kernel side.

It still works the same as before, it just works better now.

Signed-off-by: Jens Axboe <axboe@kernel.dk>