examples/proxy: expand comments on multishot and direct descriptors
Particulary for how multishot works and how the application knows
whether it will keep triggering completions or not, and on now just
why direct descriptors work, but how operations that instantiate them
(like the socket op used in proxy) can choose to either manage their
own direct descriptor space or have the kernel allocate them.

Also notes how the POSIX nonsense of mandating that the lowest free
descriptor be returned if the kernel is asked to allocate one does
not apply to io_uring. If the kernel is managing the direct descriptor
space, it may return ANY free direct descriptor without guaranteeing
it's the lowest possible one.

Signed-off-by: Jens Axboe <axboe@kernel.dk>