examples/proxy: don't just check for closed connections on empty CQ ring
We can have connections closing when we do receive CQEs. However, we
don't always want to loop and check if we need to close any. Have
anything that potentially needs that flag the connection as such.

With that, we can bump our wait timeout to 1 second. That was done
prior to this change, but then showed delays as we sometimes didn't
properly close connections in time.

Signed-off-by: Jens Axboe <axboe@kernel.dk>