examples/proxy: fix confusing mix of -r and -p for proxy vs sink mode
Sink has them reversed which is pretty annoying and ugly. With this
change, -r is always what proxy listens on, regardless of whether
it's acting as a proxy or a sink.

Update the examples as well to reflect that. Only the sink one needed
changing, but lay out the other command lines more logically too so
it's displayed as 'input side' and then 'output side'.

Signed-off-by: Jens Axboe <axboe@kernel.dk>