examples/proxy: simplify the ring buffer management and -ENOBUFS handling
We had two rings, and we'd switch between them as we hit resubmit. But
if a sender is flooding us with data and we can't keep up, then we
hit -ENOBUFS quite often. Rather than have two rings and switch between
them, it's both simpler and more efficient to just have a single ring
and wait with re-arming the receive until we've replenished a certain
amount of buffers.

We default to using half of the ring buffer size for our rearm mark.
Eg if we have 1024 buffers and we hit -ENOBUFS on a receive, then wait
until 512 sends have been processed (each which will add a buffer back
into the pool) and kick the receive back into gear at that point.

This simplifies how to deal with -ENOBUFS, and particularly the code
around having multiple struct conn_buf_ring per connection.

Signed-off-by: Jens Axboe <axboe@kernel.dk>